Cambios menores de tabulacion en cada clase, ticket mejorado, precios dinamicos hechos.

+++ CineJava/src/main/java/com/salesianostriana/dam/cinejava/repository/AjustesRepository.java
@@ -5,40 +5,40 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.salesianostriana.dam.cinejava.model.Ajustes;
 
-public interface AjustesRepository extends JpaRepository<Ajustes, Integer>{
+public interface AjustesRepository extends JpaRepository<Ajustes, Integer> {
 
 	@Query("""
-			select a.precioBaseEntrada
-			from Ajustes a
-			where a.id = :id
-		""")
+				select a.precioBaseEntrada
+				from Ajustes a
+				where a.id = :id
+			""")
 	double findPrecioBaseEntradaById(int id);
-	
+
 	@Query("""
-			select a.precioVip
-			from Ajustes a
-			where a.id = :id
-		""")
+				select a.precioVip
+				from Ajustes a
+				where a.id = :id
+			""")
 	double findPrecioVipById(int id);
-	
+
 	@Query("""
-			select a.diaDescuento
-			from Ajustes a
-			where a.id = :id
-		""")
+				select a.diaDescuento
+				from Ajustes a
+				where a.id = :id
+			""")
 	String findDiaDescuentoById(int id);
-	
+
 	@Query("""
-			select a.porcentDescuento
-			from Ajustes a
-			where a.id = :id
-		""")
+				select a.porcentDescuento
+				from Ajustes a
+				where a.id = :id
+			""")
 	double findPorcentDescuentoById(int id);
-	
+
 	@Query("""
-			select a.cantEntradasParaGratis
-			from Ajustes a
-			where a.id = :id
-		""")
-	double findCantEntradasParaGratisById(int id);
+				select a.cantEntradasParaGratis
+				from Ajustes a
+				where a.id = :id
+			""")
+	int findCantEntradasParaGratisById(int id);
 }
